refactor(recipe,recipebook): default Content() for recipe family blocs
Move all recipe-family screens into their respective public modules and
add default Content() implementations to:
- EditRecipeBookBloc, ImportRecipesBloc, ExportRecipesBloc, RecipeCategoriesBloc,
  RecipeListBloc
- EditRecipeBloc, RecipeDetailBloc, RecipeRootBloc, AddRecipeToGroceryListBloc,
  RecipePickerBloc, ChooseDateBloc, ChooseMealTypeBloc, MealPlannerRootBloc

For the recipe/core feature, screen-adjacent helpers (CategoryPickerSheet,
CropPhotoOverlay, RecipeImageFullScreenScreen) move along with their screens
so the public-side screens stay self-contained. Adds the missing
recipe.categories.public dep, util.public dep on recipe/list/public, and
updates impacted screenshot tests to point at the new public packages.
