+++ client/recipe/core/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/impl/addmeal/ChooseMealTypeBlocImpl.kt
@@ -1,7 +1,5 @@
 package com.plusmobileapps.chefmate.recipe.core.impl.addmeal
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
@@ -10,7 +8,6 @@ import com.plusmobileapps.chefmate.mapState
 import com.plusmobileapps.chefmate.meal.data.MealType
 import com.plusmobileapps.chefmate.recipe.core.addmeal.ChooseMealTypeBloc
 import com.plusmobileapps.chefmate.recipe.core.addmeal.ChooseMealTypeBloc.Output
-import com.plusmobileapps.chefmate.recipe.core.impl.addmeal.ui.ChooseMealTypeScreen
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedInject
@@ -66,10 +63,5 @@ class ChooseMealTypeBlocImpl(
 
     override fun onBackClicked() {
         output.onNext(Output.Back)
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        ChooseMealTypeScreen(bloc = this, modifier = modifier)
     }
 }
