+++ client/recipe/core/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/impl/edit/EditRecipeBlocImpl.kt
@@ -1,7 +1,5 @@
 package com.plusmobileapps.chefmate.recipe.core.impl.edit
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import chefmate.client.recipe.core.impl.generated.resources.Res
 import chefmate.client.recipe.core.impl.generated.resources.create_recipe
 import chefmate.client.recipe.core.impl.generated.resources.edit_recipe
@@ -13,7 +11,6 @@ import com.plusmobileapps.chefmate.getViewModel
 import com.plusmobileapps.chefmate.mapState
 import com.plusmobileapps.chefmate.recipe.core.edit.EditRecipeBloc
 import com.plusmobileapps.chefmate.recipe.core.edit.EditRecipeBloc.Output
-import com.plusmobileapps.chefmate.recipe.core.impl.edit.ui.EditRecipeScreen
 import com.plusmobileapps.chefmate.recipe.data.BuiltinCategory
 import com.plusmobileapps.chefmate.recipe.data.Category
 import com.plusmobileapps.chefmate.recipe.data.ExtractedRecipeData
@@ -203,10 +200,5 @@ class EditRecipeBlocImpl(
 
     override fun onBackClicked() {
         viewModel.tryToClose()
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        EditRecipeScreen(bloc = this, modifier = modifier)
     }
 }
