+++ client/recipe/core/public/build.gradle.kts
@@ -19,6 +19,7 @@ kotlin {
             implementation(projects.client.util.public)
             implementation(projects.client.grocery.data.public)
             implementation(projects.client.grocery.core.public)
+            implementation(projects.client.recipe.categories.public)
             implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(libs.arkivanov.decompose.core)
             implementation(compose.components.resources)
