+++ client/recipe/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/addgrocery/AddRecipeToGroceryListBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.recipe.core.addgrocery
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BackClickBloc
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -11,6 +13,11 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface AddRecipeToGroceryListBloc : BackClickBloc, ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        AddRecipeToGroceryListScreen(bloc = this, modifier = modifier)
+    }
 
     fun onIngredientToggled(ingredient: Int)
