+++ client/recipe/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/addmeal/ChooseMealTypeBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.recipe.core.addmeal
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BackClickBloc
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -9,6 +11,11 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface ChooseMealTypeBloc : BackClickBloc, ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        ChooseMealTypeScreen(bloc = this, modifier = modifier)
+    }
 
     fun onMealTypeSelected(mealType: MealType)
