+++ client/recipe/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/addmeal/RecipePickerScreen.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.chefmate.recipe.core.impl.addmeal.ui
+package com.plusmobileapps.chefmate.recipe.core.addmeal
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement.spacedBy
@@ -25,7 +25,6 @@ import androidx.compose.ui.unit.dp
 import chefmate.client.recipe.core.public.generated.resources.Res
 import chefmate.client.recipe.core.public.generated.resources.add_meal_plan_search_recipes
 import chefmate.client.recipe.core.public.generated.resources.add_meal_plan_select_recipe
-import com.plusmobileapps.chefmate.recipe.core.addmeal.RecipePickerBloc
 import com.plusmobileapps.chefmate.text.asTextData
 import com.plusmobileapps.chefmate.ui.components.PlusHeaderContainer
 import com.plusmobileapps.chefmate.ui.components.PlusHeaderData
