+++ client/recipe/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/core/edit/CategoryPickerSheet.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.chefmate.recipe.core.impl.edit.ui
+package com.plusmobileapps.chefmate.recipe.core.edit
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.fadeIn
