+++ client/recipe/exporter/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/exporter/ExportRecipesBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.recipe.exporter
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.arkivanov.essenty.backhandler.BackHandlerOwner
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
@@ -11,6 +13,11 @@ import kotlinx.serialization.Serializable
 
 interface ExportRecipesBloc : BackHandlerOwner, ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        ExportRecipesScreen(bloc = this, modifier = modifier)
+    }
 
     fun onRecipeToggled(id: String)
